Moved exceptions outside of ml_worker module
