+++ giskard/core/savable.py
@@ -1,5 +1,3 @@
-from typing import Generic, Optional, Set
-
 import inspect
 import logging
 import os
@@ -11,10 +9,11 @@ from pathlib import Path
 
 import cloudpickle
 import yaml
+from typing import Generic, Optional, Set
 
 from giskard.client.giskard_client import GiskardClient
 from giskard.core.core import SMT, SavableMeta
-from giskard.ml_worker.exceptions.giskard_exception import python_env_exception_helper
+from giskard.exceptions.giskard_exception import python_env_exception_helper
 from giskard.ml_worker.testing.registry.registry import tests_registry
 from giskard.settings import settings
