+++ giskard/exceptions/giskard_exception.py
